Trivial warning fixes; some are still left and are non-trivial
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@2421 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/mikmod/mmio.c
@@ -265,7 +265,7 @@ static BOOL _mm_MemReader_Read(MREADER* reader,void* ptr,size_t size)
 	s = ((MMEMREADER*)reader)->buffer;
 	s += ((MMEMREADER*)reader)->pos;
 
-	if ( ((MMEMREADER*)reader)->pos + size > ((MMEMREADER*)reader)->len) 
+	if ( ((MMEMREADER*)reader)->pos + (long)size > ((MMEMREADER*)reader)->len) 
 	{
 		((MMEMREADER*)reader)->pos = ((MMEMREADER*)reader)->len;
 		return 0; /* not enough remaining bytes */
